requirements.txt
Babel==1.3
backports.ssl-match-hostname==3.4.0.2
blinker==1.3
decorator==3.4.0
docker-py==0.7.1
Flask==0.10.1
Flask-Babel==0.9
Flask-Foundation==2.1
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.11
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pbr==0.10.7
pytz==2014.10
requests==2.4.3
six==1.9.0
speaklater==1.3
SQLAlchemy==0.9.8
sqlalchemy-migrate==0.9.4
sqlparse==0.1.14
Tempita==0.5.2
websocket-client==0.23.0
Werkzeug==0.9.6
Whoosh==2.6.0
WTForms==2.0.2
cmd2==0.6.8
termcolor==1.1.0
